+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/QualityTasksDisableKitTest.groovy
@@ -1,0 +1,77 @@
+package ru.vyarus.gradle.plugin.quality
+
+import org.gradle.testkit.runner.BuildResult
+import org.gradle.testkit.runner.TaskOutcome
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 28.08.2016
+ */
+class QualityTasksDisableKitTest extends AbstractKitTest {
+
+    def "Check java and groovy checks disable"() {
+        setup:
+        build("""
+            plugins {
+                id 'groovy'
+                id 'ru.vyarus.quality'
+            }
+
+            quality {
+                enabled = false
+            }
+
+            repositories {
+                jcenter() //required for testKit run
+            }
+
+            dependencies {
+                compile localGroovy()
+            }
+        """)
+
+        fileFromClasspath('src/main/java/sample/Sample.java', '/ru/vyarus/gradle/plugin/quality/java/sample/Sample.java')
+        fileFromClasspath('src/main/groovy/sample/GSample.groovy', '/ru/vyarus/gradle/plugin/quality/groovy/sample/GSample.groovy')
+
+        when: "run check task with both sources"
+        BuildResult result = run('check')
+
+        then: "all plugins detect violations"
+        result.task(":check").outcome == TaskOutcome.UP_TO_DATE
+        !result.output.contains('CodeNarc rule violations were found')
+        !result.output.contains('Checkstyle rule violations were found')
+        !result.output.contains('FindBugs rule violations were found')
+        !result.output.contains('PMD rule violations were found')
+    }
+
+    def "Check direct task call"() {
+        setup:
+        build("""
+            plugins {
+                id 'groovy'
+                id 'ru.vyarus.quality'
+            }
+
+            quality {
+                enabled = false
+            }
+
+            repositories {
+                jcenter() //required for testKit run
+            }
+
+            dependencies {
+                compile localGroovy()
+            }
+        """)
+
+        fileFromClasspath('src/main/java/sample/Sample.java', '/ru/vyarus/gradle/plugin/quality/java/sample/Sample.java')
+
+        when: "run check task with both sources"
+        BuildResult result = runFailed('checkstyleMain')
+
+        then: "all plugins detect violations"
+        result.task(":checkstyleMain").outcome == TaskOutcome.FAILED
+        result.output.contains('Checkstyle rule violations were found')
+    }
+}